+++ idp-commons/src/main/java/de/gematik/idp/data/fedmaster/FedmasterErrorResponse.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright (c) 2022 gematik GmbH
+ * 
+ * Licensed under the Apache License, Version 2.0 (the License);
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an 'AS IS' BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package de.gematik.idp.data.fedmaster;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder(toBuilder = true)
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class FedmasterErrorResponse {
+
+  @JsonProperty("operation")
+  private String operation;
+
+  @JsonProperty("error")
+  private String errorMessage;
+
+  @JsonProperty("error_description")
+  private String detailMessage;
+
+  @JsonProperty("gematik_timestamp")
+  private long timestamp;
+
+  @JsonProperty("gematik_uuid")
+  private String errorUuid;
+
+  @JsonProperty("gematik_code")
+  private String gematikCode;
+
+  @JsonIgnore private int httpStatusCode;
+}