+++ idp-server/src/test/java/de/gematik/idp/server/IdpControllerParameterValidationTest.java
@@ -205,7 +205,7 @@ class IdpControllerParameterValidationTest {
 
   @Test
   void getAuthenticationChallenge_emptyNonce_shouldGiveError() {
-    assertErrorResponseMatches(
+    assertForwardErrorResponseMatches(
         buildGetChallengeRequest(getInvalidationFunction("nonce", "")),
         2007,
         INVALID_REQUEST,